modify topic number selection, plot and print results
Automated selection of number of topics. Might be a bit of a mistake, as the max coherence score might be for a number of topics which is too high to be useful.

printed top 3 observations for each topic by probability score.

began plotting topics.

NEXT STEP: Merge this tidying notebook in with the scrapping notebook.